services and impl

+++ src/main/java/tn/esprit/spring/repositories/MatiereRepository.java
@@ -3,5 +3,5 @@ package tn.esprit.spring.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import tn.esprit.spring.entities.Matiere;
 
-public interface MatiereRepository extends JpaRepository<Matiere, Integer> {
+public interface MatiereRepository extends JpaRepository<Matiere, Long> {
 }
